feat: loading indicator adding to card
