+++ src/app/cart/cart-actions.interfaces.ts
@@ -1,3 +1,23 @@
+// Create cart
+export interface CartResponse {
+  id: string;
+  version: number;
+  customerId?: string;
+  anonymousId?: string;
+  lineItems: LineItem[];
+}
+
+interface LineItem {
+  id: string;
+  productId: string;
+  name: string;
+}
+
+export interface MyCartDraft {
+  currency: string;
+}
+
+// Add to cart
 export interface UpdateCart {
   version: number;
   actions: Action[];
