+++ src/app/pages/catalog-product-page/product-card/product-card.component.ts
@@ -1,15 +1,15 @@
 import { Component, computed, inject, input, OnInit } from '@angular/core';
 import { ProductProjectionResponse, ProductVariant } from '../../../products/products.interfaces';
 import { Router } from '@angular/router';
-import { NgFor, NgIf } from '@angular/common';
+import { CommonModule, NgFor, NgIf } from '@angular/common';
 import { UserDataService } from '../../../data/services/user-data.service';
 import { ROUTES_PAGES } from '../../../data/enums/routers';
-import { CreateCartService } from '../../../cart/create-cart.service';
-import { UpdateCartService } from '../../../cart/update-cart.service';
+import { BehaviorSubject } from 'rxjs';
+import { CartActionsService } from '../../../cart/cart-actions.service';
 
 @Component({
   selector: 'app-product-card',
-  imports: [NgIf, NgFor],
+  imports: [CommonModule, NgIf, NgFor],
   templateUrl: './product-card.component.html',
   styleUrl: './product-card.component.css',
 })
@@ -19,16 +19,18 @@ export class ProductCardComponent implements OnInit {
   public discountPrice: string | undefined;
   public attributes: { name: string; degree: string }[] | undefined;
   public countAttributes = 1;
+  public isAddingToCart$ = new BehaviorSubject<boolean>(false);
 
   public readonly isInCart = computed(() => {
-    const ids = this.userDataService.productIdsFromCart();
-    return ids.includes(this.product().id);
+    const productId = this.product().id?.toString();
+    const cartProductIds = this.userDataService.productIdsFromCart().map(id => id.toString());
+
+    return cartProductIds.includes(productId);
   });
 
   private router = inject(Router);
   private userDataService = inject(UserDataService);
-  private updateCartService = inject(UpdateCartService);
-  private createCartService = inject(CreateCartService);
+  private cartService = inject(CartActionsService);
 
   public ngOnInit(): void {
     this.getPrice();
@@ -36,7 +38,8 @@ export class ProductCardComponent implements OnInit {
     this.getAttributes();
   }
 
-  public addToCart(): void {
+  public addToCart(event: Event): void {
+    event.stopPropagation();
     const cart = this.userDataService.customerData()?.cart;
     const cartId = cart?.id;
     const version = cart?.version;
@@ -44,9 +47,14 @@ export class ProductCardComponent implements OnInit {
     const variantId = this.product().masterVariant.id;
 
     if (cartId && version != null) {
-      this.updateCartService.addToCard(cartId, version, productId, variantId, 1).subscribe({
+      this.isAddingToCart$.next(true);
+      this.cartService.addToCart(cartId, version, productId, variantId, 1).subscribe({
         next: () => {
           this.userDataService.refreshCustomerData();
+          this.isAddingToCart$.next(false);
+        },
+        error: () => {
+          this.isAddingToCart$.next(false);
         },
       });
     }
